+++ extensions/dashlane-vault/src/lib/dcli.ts
@@ -2,21 +2,36 @@ import { getPreferenceValues } from "@raycast/api";
 import { existsSync } from "fs";
 import { safeParse } from "valibot";
 
-import { execFilePromis } from "@/helper/exec";
+import { getMasterPassword } from "@/helper/master-password";
 import { VaultCredential, VaultCredentialSchema, VaultNote, VaultNoteSchema } from "@/types/dcli";
+import { exec, execFile } from "child_process";
+import { execa } from "execa";
+import util from "util";
 
+export const execFilePromis = util.promisify(execFile);
 const preferences = getPreferenceValues<Preferences>();
 
 const CLI_PATH =
   preferences.cliPath ?? ["/usr/local/bin/dcli", "/opt/homebrew/bin/dcli"].find((path) => existsSync(path));
 
 async function dcli(...args: string[]) {
-  if (CLI_PATH) {
-    const { stdout } = await execFilePromis(CLI_PATH, args, { maxBuffer: 4096 * 1024 });
-    return stdout;
+  if (!CLI_PATH) {
+    throw new Error("Dashlane CLI is not found!");
   }
+  const masterPassword = preferences.useTouchID ? await getMasterPassword() : null;
 
-  throw Error("Dashlane CLI is not found!");
+  const { stdout, stderr } = await execa(CLI_PATH, args, {
+    maxBuffer: 4096 * 1024,
+    env: {
+      ...(masterPassword ? { DASHLANE_MASTER_PASSWORD: masterPassword } : {}),
+    },
+  });
+  // reopen raycast as touchid closes the window
+  exec("open -a Raycast.app");
+  if (stdout.length < 0 && stderr.length > 1) {
+    throw new Error(`Failed to execute Dashlane CLI ${stderr}`);
+  }
+  return stdout;
 }
 
 export async function syncVault() {
@@ -49,6 +64,10 @@ export async function getPassword(id: string) {
 export async function getOtpSecret(id: string) {
   const stdout = await dcli("otp", `id=${id}`, "--print");
   return stdout.trim();
+}
+
+export async function encryptVault() {
+  await dcli("configure", "save-master-password", "false");
 }
 
 function parseVaultCredentials(jsonString: string): VaultCredential[] {